Fix usage of a stamped lock on L2Character#doAttack.
The return value of StampedLock#tryWriteLock must be checked to
determine if the lock could be aquired. If not, we return from
L2Character#doAttack. This also removes log spam with
IllegalMonitorStateException, where the write lock was tried to be
released with a 0 stamp which indicates that the lock was not aquired.